% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault.R
\name{backendInitialize,MsBackendMzVault-method}
\alias{backendInitialize,MsBackendMzVault-method}
\title{Initialize a \code{MsBackendMzVault}}
\usage{
\S4method{backendInitialize}{MsBackendMzVault}(object, file, implicitIsolationWidth = 1)
}
\arguments{
\item{object}{a \code{MsBackendMzVault} object}

\item{file}{the mzVault SQLite library to load}

\item{implicitIsolationWidth}{The assumed isolation width for precursor ions}
}
\description{
Initialize a \code{MsBackendMzVault}
}
