% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault.R
\name{$,MsBackendMzVault-method}
\alias{$,MsBackendMzVault-method}
\title{Access spectraData columns by name}
\usage{
\S4method{$}{MsBackendMzVault}(x, name)
}
\arguments{
\item{x}{a \code{MsBackendMzVault} object}

\item{name}{the name of the column}
}
\description{
Access spectraData columns by name
}
