% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{constant}
\alias{constant}
\title{Constant value for field mapping}
\usage{
constant(v)
}
\arguments{
\item{v}{a constant value to fill the field with}
}
\description{
Constant value for field mapping
}
