% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault.R
\name{dataStorage,MsBackendMzVault-method}
\alias{dataStorage,MsBackendMzVault-method}
\title{Report data storage to be the source SQLite database}
\usage{
\S4method{dataStorage}{MsBackendMzVault}(object)
}
\arguments{
\item{object}{a \code{MsBackendMzVault} object}
}
\description{
Report data storage to be the source SQLite database
}
