% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{fast_replace_nan}
\alias{fast_replace_nan}
\title{Replace NaN}
\usage{
fast_replace_nan(data, replace = NA)
}
\arguments{
\item{data}{the data}

\item{replace}{the replacement, typically \code{NA}}
}
\value{
data with replacement
}
\description{
Replace NaN
}
\examples{

x1 <- c(1,2,NaN,3,4)

y1 <- fast_replace_nan(x1)


}
