% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{get_db_con}
\alias{get_db_con}
\title{Get database connection}
\usage{
get_db_con(object)
}
\arguments{
\item{object}{a \code{MsBackendMzVault} object}
}
\description{
Get database connection
}
