% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{get_default_spectravariables_mapping}
\alias{get_default_spectravariables_mapping}
\title{Map spectraVariables to data from mzVault}
\usage{
get_default_spectravariables_mapping(object)
}
\arguments{
\item{object}{a MsBackendMzVault object}
}
\description{
Map spectraVariables to data from mzVault
}
