% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{get_filtered_spectrumtable}
\alias{get_filtered_spectrumtable}
\title{Get SpectrumTable filtered by specified predicates}
\usage{
get_filtered_spectrumtable(object, filters = object@filters)
}
\arguments{
\item{object}{a MsBackendMzVault object}

\item{filters}{a list of filters to filter the table by. Where are the predicates documented?}
}
\description{
Get SpectrumTable filtered by specified predicates
}
