% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{get_filtered_spectrumtable_count}
\alias{get_filtered_spectrumtable_count}
\title{Get number of rows in filtered SpectrumTable}
\usage{
get_filtered_spectrumtable_count(object)
}
\arguments{
\item{object}{a MsBackendMzVault object}
}
\description{
Get number of rows in filtered SpectrumTable
}
