% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{join_compoundtable}
\alias{join_compoundtable}
\title{Extend SpectrumTable by CompoundTable data}
\usage{
join_compoundtable(tbl_spectrum, object)
}
\arguments{
\item{tbl_spectrum}{a SpectrumTables}

\item{object}{a MsBackendMzVault object to load compound data from}
}
\description{
Extend SpectrumTable by CompoundTable data
}
