% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault.R
\name{length,MsBackendMzVault-method}
\alias{length,MsBackendMzVault-method}
\title{Get number of spectra in the filtered object}
\usage{
\S4method{length}{MsBackendMzVault}(x)
}
\arguments{
\item{x}{a \code{MsBackendMzVault} object}
}
\description{
Get number of spectra in the filtered object
}
