% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length_blob}
\alias{length_blob}
\title{Get length (in decoded elements) for a blob}
\usage{
length_blob(blob)
}
\arguments{
\item{blob}{a raw vector of bytes}
}
\value{
an integer
}
\description{
Get length (in decoded elements) for a blob
}
