% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lengths_blob}
\alias{lengths_blob}
\title{Get lengths (in decoded elements) for a list of blobs}
\usage{
lengths_blob(blobs)
}
\arguments{
\item{blobs}{a list of raw vectors of bytes}
}
\value{
a numeric vector of integers
}
\description{
Get lengths (in decoded elements) for a list of blobs
}
