% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{mzvault_read_collisionenergy}
\alias{mzvault_read_collisionenergy}
\title{Extract any number from the \code{CollisionEnergy} text field.
If multiple are found, they are averaged (e.g. for "Ramp 33.3-44-4 eV")}
\usage{
mzvault_read_collisionenergy(x)
}
\arguments{
\item{x}{a collisionEnergy field to parse}
}
\description{
Extract any number from the \code{CollisionEnergy} text field.
If multiple are found, they are averaged (e.g. for "Ramp 33.3-44-4 eV")
}
