% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{mzvault_read_polarity}
\alias{mzvault_read_polarity}
\title{Read polarity from mzVault field}
\usage{
mzvault_read_polarity(x)
}
\arguments{
\item{x}{a polarity field}
}
\description{
Read polarity from mzVault field
}
