% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault-functions.R
\name{process_filter_to_predicate}
\alias{process_filter_to_predicate}
\title{Process filter to predicates}
\usage{
process_filter_to_predicate(filter)
}
\arguments{
\item{filter}{a list of filter statements to be processed}
}
\description{
Converts the parameters set in a backend's filter
to dplyr predicates to be applied with reduce().
}
