% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_blob}
\alias{read_blob}
\title{Read a vector of doubles from raw bytes}
\usage{
read_blob(blob)
}
\arguments{
\item{blob}{a raw vector of bytes}
}
\value{
a numeric vector of doubles
}
\description{
Read a vector of doubles from raw bytes
}
