% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_blobs}
\alias{read_blobs}
\title{Read vectors of doubles from a list of raw byte vectors}
\usage{
read_blobs(blobs)
}
\arguments{
\item{blobs}{a list of raw vectors of bytes}
}
\value{
a list of numeric vectors of doubles
}
\description{
Read vectors of doubles from a list of raw byte vectors
}
