% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_blobs_numericlist}
\alias{read_blobs_numericlist}
\title{Read a NumericList from a list fo raw byte vectors}
\usage{
read_blobs_numericlist(blobs, compress = FALSE)
}
\arguments{
\item{blobs}{a list of raw vectors of bytes}

\item{compress}{a logical, should the resulting NumericList be compressed}
}
\value{
an NumericList object
}
\description{
Read a NumericList from a list fo raw byte vectors
}
