% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzVault.R
\name{[,MsBackendMzVault-method}
\alias{[,MsBackendMzVault-method}
\title{Subsetting by index}
\usage{
\S4method{[}{MsBackendMzVault}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{a \code{MsBackendMzVault} object}

\item{i}{Indices}

\item{j}{Not supported}

\item{...}{Not supported additional parameters}

\item{drop}{Not supported
@importMethodsFrom Spectra [}
}
\description{
Subsetting by index
}
