% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_blob}
\alias{write_blob}
\title{Write a vector of doubles to raw bytes}
\usage{
write_blob(data)
}
\arguments{
\item{data}{a numeric vector of doubles}
}
\value{
a raw vector of bytes
}
\description{
Write a vector of doubles to raw bytes
}
