% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_blobs}
\alias{write_blobs}
\title{Write vectors of doubles to a list of raw byte vectors}
\usage{
write_blobs(blobs)
}
\arguments{
\item{blobs}{a list of numeric vectors of doubles}
}
\value{
a list of raw vectors of bytes
}
\description{
Write vectors of doubles to a list of raw byte vectors
}
